+++ src/typer_repyt/constants.py
@@ -1,4 +1,4 @@
-from enum import Enum, Flag, auto
+from enum import Enum, Flag, IntEnum, auto
 
 
 class Sentinel(Enum):
@@ -21,3 +21,14 @@ class Validation(Flag):
     AFTER = auto()
     BOTH = BEFORE | AFTER
     NONE = 0
+
+
+class ExitCode(IntEnum):
+    """
+    Maps exit codes for the application.
+    """
+
+    SUCCESS = 0
+    GENERAL_ERROR = 1
+    CANNOT_EXECUTE = 126
+    INTERNAL_ERROR = 128