+++ src/typer_repyt/format.py
@@ -4,6 +4,7 @@ import snick
 from rich.console import Console, RenderableType
 from rich.markdown import Markdown
 from rich.panel import Panel
+from rich.text import Text
 
 
 def terminal_message(
@@ -15,6 +16,7 @@ def terminal_message(
     footer_align: Literal["left", "right", "center"] = "left",
     indent: bool = True,
     markdown: bool = False,
+    error: bool = False,
 ):
     panel_kwargs: dict[str, Any] = dict(padding=1, title_align=subject_align, subtitle_align=footer_align)
     if subject is not None:
@@ -27,20 +29,26 @@ def terminal_message(
             message = snick.indent(message, prefix="  ")
         if markdown:
             message = Markdown(message)
-    console = Console()
+    console = Console(stderr=error)
     console.print()
     console.print(Panel(message, **panel_kwargs))
     console.print()
 
 
-def simple_message(message: str, indent: bool = False, markdown: bool = False):
+def simple_message(message: str, indent: bool = False, markdown: bool = False, error: bool = False):
     text: str = snick.dedent(message)
     if indent:
         text = snick.indent(text, prefix="  ")
     content: str | Markdown = text
     if markdown:
         content = Markdown(text)
-    console = Console()
+    console = Console(stderr=error)
     console.print()
     console.print(content)
     console.print()
+
+
+def strip_rich_style(text: str | Text) -> str:
+    if isinstance(text, str):
+        text = Text.from_markup(text)
+    return text.plain